+++ jestglobalsetup.js
@@ -1,0 +1,6 @@
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+const aggregateTranslations = require('terra-toolkit/scripts/aggregate-translations/aggregate-translations');
+
+module.exports = () => {
+  aggregateTranslations();
+};