+++ packages/terra-alert/src/Alert.jsx
@@ -8,7 +8,6 @@ import IconWarning from 'terra-icon/lib/icon/IconWarning';
 import IconInformation from 'terra-icon/lib/icon/IconInformation';
 import IconSuccess from 'terra-icon/lib/icon/IconSuccess';
 import classNames from 'classnames/bind';
-import 'terra-base/lib/baseStyles';
 import styles from './Alert.module.scss';
 
 const cx = classNames.bind(styles);