+++ packages/terra-base/tests/jest/Base.test.jsx
@@ -1,13 +1,24 @@
 import React from 'react';
 import Base from '../../src/Base';
 
+// Missing locale test
+it('throws error for missing required locale', () => {
+  const messages = { Terra: 'Terra' };
+
+  try {
+    shallow(<Base customMessages={messages}>String</Base>);
+  } catch (e) {
+    expect(e.message).toContain('The prop `locale` is marked as required in `Base`');
+  }
+});
+
 // Snapshot Tests
-it('should support rendering a string without translation', () => {
+it('should support rendering a string', () => {
   const base = shallow(<Base>String</Base>);
   expect(base).toMatchSnapshot();
 });
 
-it('should support rendering an array of children without translation', () => {
+it('should support rendering an array of children', () => {
   /* eslint-disable comma-dangle */
   const base = shallow(
     <Base>
@@ -19,18 +30,7 @@ it('should support rendering an array of children without translation', () => {
   /* eslint-enable comma-dangle */
 });
 
-// Snapshot Tests
-it('should render with stict mode when strict mode enabled', () => {
+it('should render with strict mode when strict mode enabled', () => {
   const base = shallow(<Base strictMode>String</Base>);
   expect(base).toMatchSnapshot();
-});
-
-it('throws error for missing required locale', () => {
-  const messages = { Terra: 'Terra' };
-
-  try {
-    shallow(<Base customMessages={messages}>String</Base>);
-  } catch (e) {
-    expect(e.message).toContain('Missing locale prop');
-  }
 });