+++ packages/terra-form-field/src/Field.jsx
@@ -2,7 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import IconError from 'terra-icon/lib/icon/IconError';
-import 'terra-base/lib/baseStyles';
 import styles from './Field.module.scss';
 
 const cx = classNames.bind(styles);
@@ -98,6 +97,11 @@ const contextTypes = {
   },
 };
 
+const hasWhiteSpace = s => /\s/g.test(s);
+// Detect IE 10 or IE 11
+// TODO - Delete detection logic when we drop support for IE
+const isIE = () => (window.navigator.userAgent.indexOf('Trident/6.0') > -1 || window.navigator.userAgent.indexOf('Trident/7.0') > -1);
+
 const Field = (props, { intl }) => {
   const {
     children,
@@ -131,6 +135,49 @@ const Field = (props, { intl }) => {
     labelAttrs.className,
   ]);
 
+  // Checks to run when not in production
+  // TODO - Remove this once we make this a required prop.
+  // More info: https://github.com/cerner/terra-core/issues/2290
+  if (process.env.NODE_ENV !== 'production') {
+    if (!htmlFor) {
+      // eslint-disable-next-line
+      console.warn('This prop will be required in the next major version bump of terra-form-field. It is needed for creating an accessible mapping from the form field to its related error and help text.');
+    }
+
+    if (htmlFor && hasWhiteSpace(htmlFor)) {
+      // eslint-disable-next-line
+      console.warn('The htmlFor prop should be a string without white spaces as it will be used as an HTML attribute value. Use - or _ in place of white space characters.');
+    }
+  }
+
+  let helpTextId;
+  let errorTextId;
+
+  if (htmlFor) {
+    /**
+     * IE + JAWS has trouble reading aria-describedby content with our form components.
+     * In that browser, we don't want an ID for the aria-describedby content so we have a
+     * Microsoft specific feature detect to flex on if the ID is undefined or valid.
+     */
+    helpTextId = !isIE() ? `${htmlFor}-help` : undefined;
+    errorTextId = !isIE() ? `${htmlFor}-error` : undefined;
+  }
+
+  /**
+   * IE + JAWS has trouble reading aria-describedby content with our form components.
+   * Using feature detect for Microsoft browsers and injecting the help and error messages
+   * into the label as visually hidden text so JAWS can announce them correctly in IE.
+   */
+  const IEDescriptionText = (
+    isIE()
+      ? (
+        <div className={cx('visually-hidden-text')}>
+          {isInvalid && error ? error : null}
+          {help}
+        </div>
+      ) : null
+  );
+
   const labelGroup = (
     <div className={cx(['label-group', { 'label-group-hidden': isLabelHidden }])}>
       {isInvalid && <div className={cx('error-icon')}>{errorIcon}</div>}
@@ -140,6 +187,7 @@ const Field = (props, { intl }) => {
           {label}
           {required && !isInvalid && hideRequired && <div className={cx('required-hidden')}>*</div>}
           {showOptional && !required && <div className={cx('optional')}>{intl.formatMessage({ id: 'Terra.form.field.optional' })}</div>}
+          {IEDescriptionText}
         </label>
       }
       {!isInvalid && <div className={cx('error-icon-hidden')}>{errorIcon}</div>}
@@ -150,8 +198,8 @@ const Field = (props, { intl }) => {
     <div style={customStyles} {...customProps} className={fieldClasses}>
       {labelGroup}
       {children}
-      {isInvalid && error && <div className={cx('error-text')}>{error}</div>}
-      {help && <div className={cx('help-text')}>{help}</div>}
+      {isInvalid && error && <div id={errorTextId} className={cx('error-text')}>{error}</div>}
+      {help && <div id={helpTextId} className={cx('help-text')}>{help}</div>}
     </div>
   );
 };