+++ packages/terra-form-select/src/_Option.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
-import 'terra-base/lib/baseStyles';
 import SharedUtil from './_SharedUtil';
 import styles from './_Option.module.scss';