+++ packages/terra-form-select/tests/jest/Menu.test.jsx
@@ -1,14 +1,36 @@
 import React from 'react';
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+import { shallowWithIntl } from 'terra-enzyme-intl';
 import Menu from '../../src/_Menu';
 import Option from '../../src/_Option';
-import intlContexts from './intl-context-setup';
 
 describe('Menu', () => {
   it('should render a default Menu', () => {
     const liveRegion = { current: document.createElement('div') };
 
     const menu = <Menu onSelect={() => {}} visuallyHiddenComponent={liveRegion} variant="default" value="value" />;
-    const wrapper = render(menu, intlContexts.shallowContext);
+    const wrapper = shallowWithIntl(menu);
+    expect(wrapper).toMatchSnapshot();
+  });
+
+  it('should render a Menu with no results and update the live region appropriately', () => {
+    const liveRegion = { current: document.createElement('div') };
+    const mockIntl = { formatMessage: id => (`No Results for ${id.id}`) };
+
+    const menu = (
+      <Menu onSelect={() => {}} visuallyHiddenComponent={liveRegion} intl={mockIntl} variant="default" value="value" searchValue="asdf">
+        <Option value="value" display="display" />
+      </Menu>
+    );
+
+    const wrapper = shallowWithIntl(menu);
+
+    jest.useFakeTimers();
+
+    wrapper.setState({ hasNoResults: true });
+
+    jest.advanceTimersByTime(500);
+
     expect(wrapper).toMatchSnapshot();
   });
 
@@ -21,7 +43,7 @@ describe('Menu', () => {
       </Menu>
     );
 
-    const wrapper = shallow(menu, intlContexts.shallowContext);
+    const wrapper = shallowWithIntl(menu);
 
     jest.useFakeTimers();
 
@@ -41,7 +63,7 @@ describe('Menu', () => {
       </Menu>
     );
 
-    const wrapper = shallow(menu, intlContexts.shallowContext);
+    const wrapper = shallowWithIntl(menu);
 
     jest.useFakeTimers();
 
@@ -64,7 +86,7 @@ describe('Menu', () => {
 
     jest.useFakeTimers();
 
-    const wrapper = shallow(menu, intlContexts.shallowContext);
+    const wrapper = shallowWithIntl(menu);
 
     jest.runOnlyPendingTimers();