+++ packages/terra-markdown/src/Markdown.jsx
@@ -1,14 +1,35 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import classNames from 'classnames/bind';
 import marked from 'marked';
+import Prism from 'prismjs';
+import 'prismjs/components/prism-diff';
+import 'prismjs/components/prism-jsx';
+import 'prismjs/components/prism-scss';
 /* github-markdown-css's main entry in package.json resolves to a CSS file and this seems to be causing issues with eslint */
 /* eslint-disable import/extensions */
 import 'github-markdown-css/github-markdown.css';
 /* eslint-enable import/extensions */
-import './Markdown.scss';
+import styles from './Markdown.scss';
+
+const cx = classNames.bind(styles);
+
+// Create a list of loaded languages, remove functions that aren't actaully languages.
+const supportedLanguages = Object.keys(Prism.languages).filter(lang => !['extend', 'insertBefore', 'DFS'].includes(lang));
+
+const highlight = (code, lang) => {
+  let language = lang;
+  if (!supportedLanguages.includes(language)) {
+    language = 'clike';
+  }
+
+  return Prism.highlight(code, Prism.languages[language], language);
+};
 
 marked.setOptions({
   headerIds: false,
+  highlight,
+  langPrefix: 'codeblock language-',
 });
 
 const propTypes = {
@@ -26,8 +47,7 @@ const defaultProps = {
 const Markdown = props => (
   <div
     dir="ltr"
-    className="markdown-body"
-    style={{ listStyle: 'disc outside none' }}
+    className={`${cx(['markdown'])} markdown-body`}
     dangerouslySetInnerHTML={{ __html: marked(props.src) }}
   />
 );