+++ packages/terra-base/CHANGELOG.md
@@ -4,6 +4,23 @@ Changelog
 Unreleased
 ----------
 
+5.0.0 - (March 27, 2019)
+------------------
+### Breaking Changes
+* Locale prop is now required
+* This package no longer provides the `injectIntl`, `intlShape`, `FormattedMessage`, `IntlProvider` exports from react-intl
+* Removed div that wrapped children passed into the base component. The base component now uses a `<React.Fragment>` wrapper around children
+
+4.5.0 - (March 26, 2019)
+------------------
+### Changed
+* Update react/react-dom peer-dependency to latest stable minor release
+
+4.4.0 - (March 21, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 4.3.0 - (February 26, 2019)
 ------------------
 ### Changed