+++ packages/terra-doc-template/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+2.7.0 - (March 26, 2019)
+------------------
+### Changed
+* Update react/react-dom peer-dependency to latest stable minor release
+
+2.6.0 - (March 21, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.5.0 - (February 26, 2019)
 ------------------
 ### Changed