+++ packages/terra-form-field/CHANGELOG.md
@@ -4,6 +4,23 @@ ChangeLog
 Unreleased
 ----------
 
+3.6.0 - (March 26, 2019)
+------------------
+### Changed
+* Update react/react-dom peer-dependency to latest stable minor release
+
+### Removed
+* Removed peer dependency on terra-base
+* Removed baseStyles import from terra-base
+
+3.5.0 - (March 21, 2019)
+------------------
+### Added
+* Added logic to allow accessible mapping of error and help text messages when htmlFor prop is provided
+
+### Changed
+* Update Jest tests to use terra-enzyme-intl
+
 3.4.0 - (February 26, 2019)
 ------------------
 ### Changed