+++ packages/terra-form-input/CHANGELOG.md
@@ -4,6 +4,20 @@ ChangeLog
 Unreleased
 ----------
 
+2.8.0 - (March 26, 2019)
+------------------
+### Changed
+* Update react/react-dom peer-dependency to latest stable minor release
+
+### Removed
+* Removed peer dependency on terra-base
+* Removed baseStyles import from terra-base
+
+2.7.0 - (March 21, 2019)
+------------------
+### Added
+* Added logic to enable accessible mapping of InputField and it's related error and help text messages
+
 2.6.0 - (February 26, 2019)
 ------------------
 ### Changed