+++ packages/terra-list/CHANGELOG.md
@@ -3,8 +3,20 @@ Changelog
 
 Unreleased
 ----------
+
+4.3.0 - (March 26, 2019)
+------------------
 ### Changed
-* Replaced local keyCode values with keycode-js npm package 
+* Update react/react-dom peer-dependency to latest stable minor release
+
+### Removed
+* Removed peer dependency on terra-base
+* Removed baseStyles import from terra-base
+
+4.2.0 - (March 21, 2019)
+------------------
+### Changed
+* Replaced local keyCode values with keycode-js npm package
 
 4.1.0 - (February 26, 2019)
 ------------------