+++ packages/terra-props-table/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+2.28.0 - (March 26, 2019)
+------------------
+### Changed
+* Update react/react-dom peer-dependency to latest stable minor release
+
+2.27.0 - (March 21, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 2.26.0 - (February 26, 2019)
 ------------------
 ### Changed