+++ packages/terra-table/CHANGELOG.md
@@ -3,6 +3,18 @@ Changelog
 
 Unreleased
 ----------
+
+3.8.0 - (March 26, 2019)
+------------------
+### Changed
+* Update react/react-dom peer-dependency to latest stable minor release
+
+### Removed
+* Removed peer dependency on terra-base
+* Removed baseStyles import from terra-base
+
+3.7.0 - (March 21, 2019)
+------------------
 ### Changed
 * Replaced local keyCode values with keycode-js npm package